Fixed formatting and Python3 errors.

+++ mel_coefficients.py
@@ -25,8 +25,8 @@ def mel_filterbank(nfft, nfiltbank, fs):
     upper_mel = hertz_to_mel(8000)
     mel = np.linspace(lower_mel, upper_mel, nfiltbank+2)
     hertz = [mel_to_hertz(m) for m in mel]
-    fbins = [int(hz * (nfft/2+1)/fs) for hz in hertz]
-    fbank = np.empty((nfft/2+1,nfiltbank))
+    fbins = [int(hz * int(nfft/2+1)/fs) for hz in hertz]
+    fbank = np.empty((int(nfft/2+1),nfiltbank))
     for i in range(1,nfiltbank+1):
         for k in range(int(nfft/2 + 1)):
             if k < fbins[i-1]:
@@ -70,11 +70,11 @@ def mfcc(s,fs, nfiltbank):
     
     #compute periodogram
     nfft = 512
-    periodogram = np.empty((nFrames,nfft/2 + 1))
+    periodogram = np.empty((nFrames, int(nfft/2 + 1)))
     for i in range(nFrames):
         x = segment[:,i] * hamming(nSamples)
         spectrum = fftshift(fft(x,nfft))
-        periodogram[i,:] = abs(spectrum[nfft/2-1:])/nSamples
+        periodogram[i,:] = abs(spectrum[int(nfft/2-1):])/nSamples
         
     #calculating mfccs    
     fbank = mel_filterbank(nfft, nfiltbank, fs)
